requirements.txt
flask==2.0.1
python-dotenv==0.19.0
transformers==4.30.2
torch==2.0.1
numpy==1.23.5 